+++ app/src/main/java/com/example/searchanddestroy/bombscreen/ui/BombScreenUiState.kt
@@ -1,4 +1,4 @@
-package com.example.searchanddestroy.bombscreen
+package com.example.searchanddestroy.bombscreen.ui
 
 sealed class BombScreenUiState(open val password: String) {
     object Init : BombScreenUiState("")
@@ -9,6 +9,6 @@ sealed class BombScreenUiState(open val password: String) {
         override val password: String,
         val totalSeconds: Int,
         val currentMs: Float,
-        val defused : Boolean = false
+        val defused: Boolean = false
     ) : BombScreenUiState(password)
 }
